# pref: remove await/async from internal calls