--- a/src/builtin/tags/case.ts
+++ b/src/builtin/tags/case.ts
@@ -24,31 +24,17 @@ export default {
     stream.start()
   },
 
-  render: async function (ctx: Context, hash: Hash, emitter: Emitter) {
+  render: function * (ctx: Context, hash: Hash, emitter: Emitter) {
     const r = this.liquid.renderer
     for (let i = 0; i < this.cases.length; i++) {
       const branch = this.cases[i]
-      const val = await new Expression(branch.val).value(ctx)
-      const cond = await new Expression(this.cond).value(ctx)
+      const val = yield new Expression(branch.val).value(ctx)
+      const cond = yield new Expression(this.cond).value(ctx)
       if (val === cond) {
-        await r.renderTemplates(branch.templates, ctx, emitter)
+        yield r.renderTemplates(branch.templates, ctx, emitter)
         return
       }
     }
-    await r.renderTemplates(this.elseTemplates, ctx, emitter)
-  },
-
-  renderSync: function (ctx: Context, hash: Hash, emitter: Emitter) {
-    const r = this.liquid.renderer
-    for (let i = 0; i < this.cases.length; i++) {
-      const branch = this.cases[i]
-      const val = new Expression(branch.val).valueSync(ctx)
-      const cond = new Expression(this.cond).valueSync(ctx)
-      if (val === cond) {
-        r.renderTemplatesSync(branch.templates, ctx, emitter)
-        return
-      }
-    }
-    r.renderTemplatesSync(this.elseTemplates, ctx, emitter)
+    yield r.renderTemplates(this.elseTemplates, ctx, emitter)
   }
 } as ITagImplOptions