--- a/src/builtin/tags/continue.ts
+++ b/src/builtin/tags/continue.ts
@@ -1,7 +1,7 @@
 import { Emitter, Context, Hash } from '../../types'
 
 export default {
-  render: async function (ctx: Context, hash: Hash, emitter: Emitter) {
+  render: function (ctx: Context, hash: Hash, emitter: Emitter) {
     emitter.continue = true
   }
 }