--- a/src/builtin/tags/layout.ts
+++ b/src/builtin/tags/layout.ts
@@ -19,12 +19,9 @@ export default {
 
     this.tpls = this.liquid.parser.parse(remainTokens)
   },
-  render: async function (ctx: Context, hash: Hash, emitter: Emitter) {
+  render: function * (ctx: Context, hash: Hash, emitter: Emitter) {
     const layout = ctx.opts.dynamicPartials
-      ? (ctx.sync
-        ? new Expression(this.layout).valueSync(ctx)
-        : await new Expression(this.layout).value(ctx)
-      )
+      ? yield new Expression(this.layout).value(ctx)
       : this.staticLayout
     assert(layout, `cannot apply layout with empty filename`)
 
@@ -32,20 +29,14 @@ export default {
     ctx.setRegister('blockMode', BlockMode.STORE)
     const blocks = ctx.getRegister('blocks')
     const r = this.liquid.renderer
-    const html = ctx.sync
-      ? r.renderTemplatesSync(this.tpls, ctx)
-      : await r.renderTemplates(this.tpls, ctx)
+    const html = yield r.renderTemplates(this.tpls, ctx)
     if (blocks[''] === undefined) {
       blocks[''] = html
     }
-    const templates = ctx.sync
-      ? this.liquid.parseFileSync(layout, ctx.opts)
-      : await this.liquid.parseFile(layout, ctx.opts)
+    const templates = yield this.liquid._parseFile(layout, ctx.opts, ctx.sync)
     ctx.push(hash)
     ctx.setRegister('blockMode', BlockMode.OUTPUT)
-    const partial = ctx.sync
-      ? r.renderTemplatesSync(templates, ctx)
-      : await r.renderTemplates(templates, ctx)
+    const partial = yield r.renderTemplates(templates, ctx)
     ctx.pop()
     emitter.write(partial)
   }