--- a/src/builtin/tags/raw.ts
+++ b/src/builtin/tags/raw.ts
@@ -1,4 +1,4 @@
-import { TagToken, Token, ITagImplOptions, Context } from '../../types'
+import { Hash, Emitter, TagToken, Token, ITagImplOptions, Context } from '../../types'
 
 export default {
   parse: function (tagToken: TagToken, remainTokens: Token[]) {
@@ -15,7 +15,7 @@ export default {
       })
     stream.start()
   },
-  render: function (ctx: Context) {
-    return this.tokens.map((token: Token) => token.raw).join('')
+  render: function (ctx: Context, hash: Hash, emitter: Emitter) {
+    emitter.write(this.tokens.map((token: Token) => token.raw).join(''))
   }
 } as ITagImplOptions