--- a/src/template/filter/filter.ts
+++ b/src/template/filter/filter.ts
@@ -21,19 +21,11 @@ export class Filter {
     this.impl = impl || (x => x)
     this.args = args
   }
-  public async render (value: any, context: Context) {
+  public * render (value: any, context: Context) {
     const argv: any[] = []
     for (const arg of this.args) {
-      if (isKeyValuePair(arg)) argv.push([arg[0], await new Expression(arg[1]).evaluate(context)])
-      else argv.push(await new Expression(arg).evaluate(context))
-    }
-    return this.impl.apply({ context }, [value, ...argv])
-  }
-  public renderSync (value: any, context: Context) {
-    const argv: any[] = []
-    for (const arg of this.args) {
-      if (isKeyValuePair(arg)) argv.push([arg[0], new Expression(arg[1]).evaluateSync(context)])
-      else argv.push(new Expression(arg).evaluateSync(context))
+      if (isKeyValuePair(arg)) argv.push([arg[0], yield new Expression(arg[1]).evaluate(context)])
+      else argv.push(yield new Expression(arg).evaluate(context))
     }
     return this.impl.apply({ context }, [value, ...argv])
   }