--- a/src/template/itemplate.ts
+++ b/src/template/itemplate.ts
@@ -4,6 +4,5 @@ import { Emitter } from '../render/emitter'
 
 export interface ITemplate {
   token: Token;
-  render(ctx: Context, emitter: Emitter): Promise<any>;
-  renderSync(ctx: Context, emitter: Emitter): any;
+  render(ctx: Context, emitter: Emitter): any;
 }