@harttle 2019-10-26
:tada: This PR is included in version 9.2.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/liquidjs)
- [GitHub release](https://github.com/harttle/liquidjs/releases/tag/v9.2.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: